Improve Translator Switching
Rather than a bunch of if statements, we now have a dict mapping
keywords to the appropriate translator. Having this mapping in the same
place we set up HTTP routing resonates with me.

To accomplish this, we also modified handlers so that translators are
injected via the constructor (in the given mapping form) rather than
instantiating them directly. This pattern should make potential
unit-tests easier to write down the road.

Implements issue #1